+++ index.html
@@ -235,6 +235,7 @@
     </section>
   </template>
   <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module" defer></script>
 </body>
 
 </html>
